+++ app/src/main/java/de/christian2003/petrolindex/view/petrol_entries/PetrolEntriesView.kt
@@ -12,6 +12,7 @@ import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.lazy.LazyColumn
 import androidx.compose.foundation.lazy.items
+import androidx.compose.material3.ButtonDefaults
 import androidx.compose.material3.CenterAlignedTopAppBar
 import androidx.compose.material3.ExperimentalMaterial3Api
 import androidx.compose.material3.Icon
@@ -291,7 +292,8 @@ fun ConfirmDeleteDialog(
                         }.invokeOnCompletion {
                             onConfirmDelete(petrolEntry)
                         }
-                    }
+                    },
+                    colors = ButtonDefaults.textButtonColors().copy(contentColor = MaterialTheme.colorScheme.error)
                 ) {
                     Text(text = stringResource(R.string.petrol_entries_delete_confirm))
                 }